gateway-data-*: rename and make EntityManagerProducer not a utility class
