+++ gateway-data-deltaspike/src/main/java/net/kemitix/naolo/gateway/data/deltaspike/EntityManagerProducer.java
@@ -19,11 +19,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.kemitix.naolo.gateway.data.jpa;
+package net.kemitix.naolo.gateway.data.deltaspike;
 
 import lombok.extern.log4j.Log4j2;
 
 import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Disposes;
 import javax.enterprise.inject.Produces;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
@@ -35,11 +36,9 @@ import javax.persistence.Persistence;
  * @author Paul Campbell (pcampbell@kemitix.net)
  */
 @Log4j2
-@ApplicationScoped
-@SuppressWarnings("hideutilityclassconstructor")
-public class EntityManagerHelper {
+public class EntityManagerProducer {
 
-    private static final String UNIT_NAME = EntityManagerHelper.class.getPackage().getName();
+    private static final String UNIT_NAME = EntityManagerProducer.class.getPackage().getName();
 
     /**
      * Producer for EntityManagerFactory.
@@ -48,7 +47,7 @@ public class EntityManagerHelper {
      */
     @Produces
     @ApplicationScoped
-    public static EntityManagerFactory entityManagerFactory() {
+    public EntityManagerFactory entityManagerFactory() {
         log.info("Create EntityManagerFactory");
         return Persistence.createEntityManagerFactory(UNIT_NAME);
     }
@@ -60,9 +59,18 @@ public class EntityManagerHelper {
      * @return an EntityManager
      */
     @Produces
-    public static EntityManager entityManager(final EntityManagerFactory entityManagerFactory) {
+    public EntityManager entityManager(final EntityManagerFactory entityManagerFactory) {
         log.info("Create EntityManager");
         return entityManagerFactory.createEntityManager();
     }
 
+    /**
+     * Closes the EntityManager.
+     *
+     * @param em the EntityManager to close
+     */
+    public void close(@Disposes final EntityManager em) {
+        log.info("Close EntityManager");
+        em.close();
+    }
 }
